requirements.txt
Django==2.2.2
Sphinx==2.1.2
Pillow==6.0.0
django-tinymce==2.8.0